requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
langchain
langchain-openai
langchain-community
faiss-cpu